@@ -9,11 +9,11 @@
     "@types/d3": "5.7.2",
     "@types/jest": "24.0.18",
     "@types/lodash": "4.14.138",
-    "@types/node": "12.7.2",
+    "@types/node": "12.7.3",
     "@types/ramda": "0.26.19",
-    "@types/react": "16.8.25",
+    "@types/react": "16.9.2",
     "@types/react-autosuggest": "9.3.10",
-    "@types/react-dom": "16.8.5",
+    "@types/react-dom": "16.9.0",
     "autosuggest-highlight": "^3.1.1",
     "classnames": "^2.2.6",
     "cytoscape": "3.9.1",
@@ -25,7 +25,7 @@
     "cytoscape-spread": "3.0.0",
     "d3": "5.11.0",
     "dagre": "0.8.4",
-    "date-fns": "2.0.0-alpha.7",
+    "date-fns": "2.0.1",
     "file-saver": "2.0.2",
     "immutable": "^3.8.2",
     "klayjs": "^0.4.1",
@@ -35,24 +35,24 @@
     "mobx-react": "6.1.3",
     "prop-types": "15.7.2",
     "ramda": "0.26.1",
-    "react": "16.8.6",
+    "react": "16.9.0",
     "react-autosuggest": "9.4.3",
     "react-day-picker": "^7.3.2",
-    "react-dom": "16.8.6",
+    "react-dom": "16.9.0",
     "react-router-dom": "5.0.1",
-    "react-scripts": "3.0.1",
+    "react-scripts": "3.1.1",
     "react-shortcuts": "^2.0.1",
     "recompose": "0.30.0",
     "seamless-immutable": "7.1.4",
     "typeface-roboto": "^0.0.75",
-    "typescript": "3.5.3",
+    "typescript": "3.6.2",
     "wretch": "1.5.4"
   },
   "devDependencies": {
     "@storybook/react": "5.1.11",
     "@types/recompose": "0.30.7",
     "eslint-config-prettier": "6.1.0",
-    "husky": "3.0.4",
+    "husky": "3.0.5",
     "jest-junit": "7.0.0",
     "prettier": "1.18.2",
     "pretty-quick": "1.11.1",
